Passing the header with the token so that it can retrieve the data from the api

+++ AddressApp/AddressAppServer/AddressAppServer.Web/Services/AddressClient.cs
@@ -1,6 +1,7 @@
 ﻿using AddressAppServer.ClassLibrary.Common;
 using AddressAppServer.ClassLibrary.DTOs;
 using AddressAppServer.Web.Services.Interfaces;
+using Microsoft.AspNetCore.Components.Server.ProtectedBrowserStorage;
 using System.Text;
 using System.Text.Json;
 
@@ -10,32 +11,55 @@ namespace AddressAppServer.Web.Services
     {
         private readonly HttpClient _httpClient;
         private readonly ILogger<AddressClient> _logger;
+        private readonly ProtectedSessionStorage _protectedSessionStorage;
 
-        public AddressClient(HttpClient httpClient, ILogger<AddressClient> logger)
+        public AddressClient(HttpClient httpClient, ILogger<AddressClient> logger, ProtectedSessionStorage protectedSessionStorage)
         {
             _httpClient = httpClient;
             _logger = logger;
+            _protectedSessionStorage = protectedSessionStorage;
+        }
+
+        private async Task<bool> SetAuthorizationHeaderAsync()
+        {
+            var tokenResult = await _protectedSessionStorage.GetAsync<string>("authToken");
+            if (tokenResult.Success && tokenResult.Value != null)
+            {
+                _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", tokenResult.Value);
+                return true;
+            }
+            return false;
         }
 
         public async Task<Result<GetAddressesResponseDTO>> GetAddresses()
         {
             Result<GetAddressesResponseDTO> result = new();
-
-            var response = await _httpClient.GetAsync("api/Addresses");
-            string? content = await response.Content.ReadAsStringAsync();
-            result = JsonSerializer.Deserialize<Result<GetAddressesResponseDTO>>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
-
+            if (await SetAuthorizationHeaderAsync())
+            {
+                var response = await _httpClient.GetAsync("api/Addresses");
+                string? content = await response.Content.ReadAsStringAsync();
+                result = JsonSerializer.Deserialize<Result<GetAddressesResponseDTO>>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
+            }
+            else
+            {
+                result.Message = "Authentication token not found.";
+            }
             return result;
         }
 
         public async Task<Result<GetAddressResponseDTO>> GetAddress(Guid id)
         {
             Result<GetAddressResponseDTO> result = new();
-
-            var response = await _httpClient.GetAsync($"api/addresses/{id}");
-            string? content = await response.Content.ReadAsStringAsync();
-            result = JsonSerializer.Deserialize<Result<GetAddressResponseDTO>>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
-
+            if (await SetAuthorizationHeaderAsync())
+            {
+                var response = await _httpClient.GetAsync($"api/addresses/{id}");
+                string? content = await response.Content.ReadAsStringAsync();
+                result = JsonSerializer.Deserialize<Result<GetAddressResponseDTO>>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
+            }
+            else
+            {
+                result.Message = "Authentication token not found.";
+            }
             return result;
         }
 
@@ -45,10 +69,16 @@ namespace AddressAppServer.Web.Services
             string jsonPayload = JsonSerializer.Serialize(requestDTO);
             StringContent? requestContent = new StringContent(jsonPayload, Encoding.UTF8, "application/json");
 
-            using HttpResponseMessage response = await _httpClient.PostAsync("api/addresses", requestContent);
-            string? content = await response.Content.ReadAsStringAsync();
-            result = JsonSerializer.Deserialize<Result>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
-
+            if (await SetAuthorizationHeaderAsync())
+            {
+                using HttpResponseMessage response = await _httpClient.PostAsync("api/addresses", requestContent);
+                string? content = await response.Content.ReadAsStringAsync();
+                result = JsonSerializer.Deserialize<Result>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
+            }
+            else
+            {
+                result.Message = "Authentication token not found.";
+            }
             return result;
         }
 
@@ -58,21 +88,32 @@ namespace AddressAppServer.Web.Services
             string jsonPayload = JsonSerializer.Serialize(requestDTO);
             StringContent? requestContent = new StringContent(jsonPayload, Encoding.UTF8, "application/json");
 
-            using HttpResponseMessage response = await _httpClient.PutAsync("api/addresses", requestContent);
-            string? content = await response.Content.ReadAsStringAsync();
-            result = JsonSerializer.Deserialize<Result>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
-
+            if (await SetAuthorizationHeaderAsync())
+            {
+                using HttpResponseMessage response = await _httpClient.PutAsync("api/addresses", requestContent);
+                string? content = await response.Content.ReadAsStringAsync();
+                result = JsonSerializer.Deserialize<Result>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
+            }
+            else
+            {
+                result.Message = "Authentication token not found.";
+            }
             return result;
         }
 
         public async Task<Result> DeleteAddress(Guid id)
         {
             Result result = new();
-
-            var response = await _httpClient.DeleteAsync($"api/addresses/{id}");
-            string? content = await response.Content.ReadAsStringAsync();
-            result = JsonSerializer.Deserialize<Result>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
-
+            if (await SetAuthorizationHeaderAsync())
+            {
+                var response = await _httpClient.DeleteAsync($"api/addresses/{id}");
+                string? content = await response.Content.ReadAsStringAsync();
+                result = JsonSerializer.Deserialize<Result>(content, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) ?? new();
+            }
+            else
+            {
+                result.Message = "Authentication token not found.";
+            }
             return result;
         }
     }
